Migrate .sh files to "real" shell scripts
Modify (almost) all "shell" so that they run autonomously
- Make them executable
- Prepend them with a "shebang" call to execute bash
- In the composite scripts call them as subshell instead of sourcing

This enables to run each of the steps autonomously, e.g., as module
in the composite scripts (all*.sh), from the shell, from the IDE (tested
with IntelliJ).

Note that this was only tested on macOS so far.
Test on other OSs (Linux, Windows) will soon be tested via GH actions.
Besides this, some special issues (like migration from JDK 14 to 17)
will be addressed later.

Introduced in the course of support-and-care/maven-support-and-care#137